a bit more hook cleanup
