+++ app/src/organisms/ErrorRecoveryFlows/hooks/index.ts
@@ -4,10 +4,9 @@ export { useErrorMessage } from './useErrorMessage'
 export { useErrorName } from './useErrorName'
 export { useRecoveryCommands } from './useRecoveryCommands'
 export { useRouteUpdateActions } from './useRouteUpdateActions'
-export { useRecoveryTipStatus } from './useRecoveryTipStatus'
-export { usePreviousRecoveryRoute } from './usePreviousRecoveryRoute'
-export { useRecoveryRouting } from './useRecoveryRouting'
+export { useERUtils } from './useERUtils'
 
 export type { UseRouteUpdateActionsResult } from './useRouteUpdateActions'
 export type { UseRecoveryCommandsResult } from './useRecoveryCommands'
 export type { RecoveryTipStatusUtils } from './useRecoveryTipStatus'
+export type { ERUtilsResults } from './useERUtils'
